feat(gui): add GUI using Kivy
This is a simple GUI, it now let the user see the folder it picked, and when the images are going to be saved. There is an Open Folder. button, and a Start button, and some labels to indicate when everything is completed.
For this simple GUI we use Kivy, pliers and pyobjus