bug - boolean check on loadresourcepath and onEnterchangekey
pIsProduction paramete was changed by a global constant at the beginig of the script, this because on production we had to change that boolean on every funciton call, now we just need to change the consant to TRUE when in production.

Also we have a couple more checks when changing the key, for instance when the user wont input any key,  o just whitespaces. If that happens the popup will not go away until the user inputs some text, or when the user clicks outside or hit esc